cleanup iOS extensions
